requirements.txt
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.9.0
fonttools==4.44.0
fsspec==2023.4.0
idna==3.4
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==9.3.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.1
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.9
